Move OEQThumb component into regular components directory, since it makes more sense for it to live there.
