+++ Source/Plugins/Core/com.equella.core/js/tsrc/components/OEQThumb.tsx
